// gen/ts/open_splunk/knowledge.ts
export enum KnowledgeObjectType {
  KNOWLEDGE_OBJECT_TYPE_UNSPECIFIED = 0,
  KNOWLEDGE_OBJECT_TYPE_FIELD_EXTRACTION = 1,
  KNOWLEDGE_OBJECT_TYPE_FIELD_ALIAS = 2,
  KNOWLEDGE_OBJECT_TYPE_CALCULATED_FIELD = 3,
  UNRECOGNIZED = -1,
}